+++ tests/fixtures/probes.py
@@ -5,120 +5,152 @@ import sys
 EmptyProbe = {}
 
 MissingTypeProbe = {
-    "module": "blah",
-    "title": "a title",
-    "layer": "kubernetes"
+    "name": "a name",
+    "provider": {
+        "module": "blah"
+    }
 }
 
 UnknownTypeProbe = {
     "type": "whatever",
-    "title": "a title",
-    "layer": "kubernetes"
+    "name": "a name",
+    "provider": {
+        "type": "python"
+    }
+}
+
+UnknownProviderTypeProbe = {
+    "type": "probe",
+    "name": "a name",
+    "provider": {
+        "type": "pizza"
+    }
 }
 
 MissingModuleProbe = {
-    "type": "python",
-    "title": "a title",
-    "layer": "kubernetes"
+    "type": "probe",
+    "name": "a name",
+    "provider": {
+        "type": "python"
+    }
 }
 
 NotImportableModuleProbe = {
-    "type": "python",
-    "module": "fake.module",
-    "func": "myfunc",
-    "title": "a title",
-    "layer": "kubernetes"
+    "type": "probe",
+    "name": "a name",
+    "provider": {
+        "type": "python",
+        "module": "fake.module",
+        "func": "myfunc"
+    }
 }
 
 MissingFunctionProbe = {
-    "type": "python",
-    "module": "os.path",
-    "title": "missing function name",
-    "title": "a title",
-    "layer": "kubernetes"
+    "type": "probe",
+    "provider": {
+        "type": "python",
+        "module": "os.path"
+    },
+    "name": "a name"
 }
 
 MissingProcessPathProbe = {
-    "type": "process",
-    "title": "missing proc path",
-    "layer": "kubernetes"
+    "type": "probe",
+    "provider": {
+        "type": "process"
+    },
+    "name": "missing proc path"
 }
 
 ProcessPathDoesNotExistProbe = {
-    "type": "process",
-    "path": "somewhere/not/here",
-    "title": "invalid proc path",
-    "layer": "kubernetes"
+    "type": "probe",
+    "provider": {
+        "type": "process",
+        "path": "somewhere/not/here",
+    },
+    "name": "invalid proc path"
 }
 
 MissingHTTPUrlProbe = {
-    "type": "http",
-    "title": "A probe without url",
-    "layer": "kubernetes"
+    "type": "probe",
+    "provider": {
+        "type": "http"
+    },
+    "name": "A probe without url"
 }
 
 MissingFuncArgProbe = {
-    "type": "python",
-    "module": "os.path",
-    "func": "exists",
-    "title": "a title",
-    "layer": "kubernetes",
-    "arguments": {}
+    "type": "probe",
+    "provider": {
+        "type": "python",
+        "module": "os.path",
+        "func": "exists",
+        "arguments": {}
+    },
+    "name": "a name"
 }
 
 TooManyFuncArgsProbe = {
-    "title": "This probe has way too many args",
-    "type": "python",
-    "module": "os.path",
-    "func": "exists",
-    "layer": "kubernetes",
-    "arguments": {
-        "path": "/some/path",
-        "should_not_be_here": "indeed not"
+    "type": "probe",
+    "name": "This probe has way too many args",
+    "provider": {
+        "type": "python",
+        "module": "os.path",
+        "func": "exists",
+        "arguments": {
+            "path": "/some/path",
+            "should_not_be_here": "indeed not"
+        }
     }
 }
 
 PythonModuleProbe = {
-    "title": "This probe is a Python probe",
-    "type": "python",
-    "module": "os.path",
-    "func": "exists",
-    "layer": "kubernetes",
-    "arguments": {
-        "path": __file__,
-    },
-    "timeout": 30,
+    "type": "probe",
+    "name": "This probe is a Python probe",
     "pauses": {
         "before": 0,
         "after": 0.1
+    },
+    "provider": {
+        "type": "python",
+        "module": "os.path",
+        "func": "exists",
+        "arguments": {
+            "path": __file__,
+        },
+        "timeout": 30
     }
 }
 
 ProcProbe = {
-    "title": "This probe is a process probe",
-    "type": "process",
-    "path": sys.executable,
-    "layer": "kubernetes",
-    "arguments": {
-        "-V": None
-    },
-    "timeout": 1,
+    "type": "probe",
+    "name": "This probe is a process probe",
     "pauses": {
         "before": 0,
         "after": 0.1
+    },
+    "provider": {
+        "type": "process",
+        "path": sys.executable,
+        "arguments": {
+            "-V": None
+        },
+        "timeout": 1
     }
 }
 
 HTTPProbe = {
-    "title": "This probe is a HTTP probe",
-    "type": "http",
-    "url": "http://example.com",
-    "method": "post",
-    "layer": "kubernetes",
-    "arguments": {
-        "q": "chaostoolkit",
+    "type": "probe",
+    "name": "This probe is a HTTP probe",
+    "provider": {
+        "type": "http",
+        "url": "http://example.com",
+        "method": "post",
+        "arguments": {
+            "q": "chaostoolkit",
+        },
+        "timeout": 30
     },
-    "timeout": 30,
     "pauses": {
         "before": 0,
         "after": 0.1
@@ -126,13 +158,15 @@ HTTPProbe = {
 }
 
 BackgroundPythonModuleProbe = {
-    "title": "This probe is a Python probe",
-    "type": "python",
-    "module": "os.path",
-    "func": "exists",
-    "layer": "kubernetes",
-    "arguments": {
-        "path": __file__,
-    },
-    "background": True
+    "type": "probe",
+    "name": "This probe is a Python probe",
+    "background": True,
+    "provider": {
+        "type": "python",
+        "module": "os.path",
+        "func": "exists",
+        "arguments": {
+            "path": __file__,
+        }
+    }
 }